chore: 메인 페이지 영역 분리
