+++ src/pages/home/MainPage.tsx
@@ -1,15 +1,12 @@
 import styled from '@emotion/styled'
 import { ComponentProps } from 'react'
 
-import Footprint from '@/assets/icons/Footprint.tsx'
-import CubeButton from '@/components/common/cubeButton'
-import { FlexBox } from '@/components/common/flexBox'
 import Spacing from '@/components/common/spacing'
 import { Text } from '@/components/common/text'
-import Carousel from '@/pages/home/carousel'
+import ChannelList from '@/pages/home/channelList/index.tsx'
+import TodayChannelSlider from '@/pages/home/channelSlider'
+import InterestHeader from '@/pages/home/interestHeader/index.tsx'
 import { KeyOfPalette, KeyOfTypo } from '@/styles/theme'
-
-import { channelMock } from '../../mock/channel.ts'
 
 interface MainPageProps extends ComponentProps<'div'> {
   TodayChannelTypo?: KeyOfTypo
@@ -25,97 +22,27 @@ const MainPage = ({
   InterestingChannelColor = 'BLACK',
   ...props
 }: MainPageProps) => {
-  const carouselItems = [
-    'https://picsum.photos/480/190/?random',
-    'https://picsum.photos/400/190/?random',
-    'https://picsum.photos/350/190/?random',
-    'https://picsum.photos/380/190/?random',
-  ]
   return (
-    <>
+    <MainPageWrapper {...props}>
       <TodayChannel>
         <Text typo={TodayChannelTypo} color={TodayChannelColor}>
           {'오늘의 채널'}
         </Text>
         <Spacing size={15}></Spacing>
-        <TodayChannelSlider>
-          <Carousel images={carouselItems} />
-        </TodayChannelSlider>
+        <TodayChannelSlider />
       </TodayChannel>
       <Spacing size={40}></Spacing>
-      <InterestingChannel>
-        <FlexBox justify={'flex-start'} gap={10}>
-          <StyledHeader typo={InterestingChannelTypo} color={InterestingChannelColor}>
-            {'관심사에 맞는 채널을 탐색해보세요.'}
-          </StyledHeader>
-          <IconWrapper>
-            <FirstIcon>
-              <Footprint />
-            </FirstIcon>
-            <SecondIcon>
-              <Footprint />
-            </SecondIcon>
-          </IconWrapper>
-        </FlexBox>
+      <InterestChannel>
+        <InterestHeader typo={InterestingChannelTypo} color={InterestingChannelColor} />
         <Spacing size={30}></Spacing>
-        <ChannelWrapper>
-          {channelMock &&
-            channelMock.map((channel, index) => {
-              if (index % 2 === 0) {
-                const nextChannel = channelMock[index + 1]
-
-                return (
-                  <div key={channel._id}>
-                    <FlexBox gap={13}>
-                      <CubeButton content={channel.name} path={`/${channel._id}`}></CubeButton>
-                      {nextChannel && (
-                        <CubeButton
-                          content={nextChannel.name}
-                          path={`/${nextChannel._id}`}
-                        ></CubeButton>
-                      )}
-                    </FlexBox>
-                    <Spacing size={13}></Spacing>
-                  </div>
-                )
-              }
-
-              return null
-            })}
-        </ChannelWrapper>
-      </InterestingChannel>
-    </>
+        <ChannelList />
+      </InterestChannel>
+    </MainPageWrapper>
   )
 }
 
+const MainPageWrapper = styled.div``
 const TodayChannel = styled.div``
-const InterestingChannel = styled.div``
-const TodayChannelSlider = styled.div`
-  width: 100%; /* 너비를 100%로 설정하여 부모 요소에 맞게 확장됩니다. */
-  height: 190px;
-  overflow: hidden;
-  border-radius: 10px;
-`
-const ChannelWrapper = styled.div``
-
-const StyledHeader = styled(Text)`
-  width: 55%;
-  line-height: 1.3;
-  word-break: keep-all;
-`
-
-const IconWrapper = styled.div`
-  position: relative;
-`
-const FirstIcon = styled.div`
-  position: absolute;
-  bottom: 2px;
-`
-
-const SecondIcon = styled.div`
-  position: absolute;
-  left: 20px;
-  top: 2px;
-`
+const InterestChannel = styled.div``
 
 export default MainPage
